Add ArtifactWorkbench contract and adapters
