+++ contracts/gateway/v4/workbench/artifacts-list.schema.json
@@ -1,0 +1,139 @@
+{
+  "$schema": "https://json-schema.org/draft/2020-12/schema",
+  "$id": "https://opensquilla.dev/contracts/gateway/v4/workbench/artifacts-list.schema.json",
+  "title": "OpenSquilla artifacts.list Contract",
+  "description": "Session-scoped immutable artifact catalog used by the ArtifactWorkbench Adapter.",
+  "type": "object",
+  "additionalProperties": false,
+  "properties": {
+    "request": { "$ref": "#/$defs/ArtifactsListRequestFrame" },
+    "response": { "$ref": "#/$defs/ArtifactsListResponseFrame" }
+  },
+  "x-opensquilla-wire": {
+    "protocol": "opensquilla-websocket-json",
+    "version": 4,
+    "compatibility": "exact-json-tree"
+  },
+  "x-opensquilla-codegen": {
+    "python": { "tool": "datamodel-code-generator", "version": "0.75.1", "target": "pydantic_v2.BaseModel" },
+    "typescript": { "tool": "json-schema-to-typescript", "version": "15.0.4" },
+    "runtimeValidation": { "tool": "ajv", "version": "8.17.1", "mode": "standalone-adapter-only" }
+  },
+  "x-opensquilla-method": {
+    "name": "artifacts.list",
+    "kind": "query",
+    "scope": "operator.read",
+    "guestAllowed": false,
+    "idempotency": "read-only",
+    "timeout": { "policy": "caller" },
+    "capability": { "kind": "method-availability", "name": "artifacts.list" },
+    "request": "#/$defs/ArtifactsListRequestFrame",
+    "params": "#/$defs/ArtifactsListParams",
+    "response": "#/$defs/ArtifactsListResponseFrame",
+    "result": "#/$defs/ArtifactsListResult",
+    "errors": [
+      { "code": "INVALID_REQUEST" },
+      { "code": "UNAUTHORIZED" },
+      { "code": "UNAVAILABLE", "retryable": true },
+      { "code": "INTERNAL_ERROR" }
+    ]
+  },
+  "$defs": {
+    "ArtifactsListParams": {
+      "type": "object",
+      "additionalProperties": true,
+      "required": ["sessionKey"],
+      "properties": {
+        "sessionKey": { "type": "string", "minLength": 1 },
+        "limit": { "type": ["integer", "string", "boolean", "null"] },
+        "before": { "type": ["string", "null"] }
+      }
+    },
+    "Artifact": {
+      "type": "object",
+      "additionalProperties": true,
+      "required": ["id", "name", "mime", "size", "download_url"],
+      "properties": {
+        "id": { "type": "string", "minLength": 1 },
+        "kind": { "type": ["string", "null"] },
+        "sha256": { "type": ["string", "null"] },
+        "name": { "type": "string" },
+        "mime": { "type": "string" },
+        "size": { "type": "integer", "minimum": 0 },
+        "session_id": { "type": ["string", "null"] },
+        "source": { "type": ["string", "null"] },
+        "created_at": { "type": ["string", "integer", "null"] },
+        "store": { "type": ["string", "null"] },
+        "download_url": { "type": "string", "minLength": 1 },
+        "thumbnail_url": { "type": ["string", "null"] }
+      }
+    },
+    "ArtifactsListResult": {
+      "type": "object",
+      "additionalProperties": false,
+      "required": ["artifacts", "has_more", "oldest_cursor", "newest_cursor", "total_count", "page_size"],
+      "properties": {
+        "artifacts": { "type": "array", "items": { "$ref": "#/$defs/Artifact" } },
+        "has_more": { "type": "boolean" },
+        "oldest_cursor": { "type": ["string", "null"] },
+        "newest_cursor": { "type": ["string", "null"] },
+        "total_count": { "type": "integer", "minimum": 0 },
+        "page_size": { "type": "integer", "minimum": 1, "maximum": 200 }
+      }
+    },
+    "RpcError": {
+      "type": "object",
+      "additionalProperties": true,
+      "required": ["code", "message"],
+      "properties": {
+        "code": { "type": "string" },
+        "message": { "type": "string" },
+        "details": {},
+        "retryable": { "type": ["boolean", "null"] },
+        "retry_after_ms": { "type": ["integer", "null"] },
+        "accepted": { "type": ["boolean", "null"] }
+      }
+    },
+    "ArtifactsListRequestFrame": {
+      "type": "object",
+      "additionalProperties": false,
+      "required": ["type", "id", "method", "params"],
+      "properties": {
+        "type": { "const": "req" },
+        "id": { "type": "string" },
+        "method": { "const": "artifacts.list" },
+        "params": { "$ref": "#/$defs/ArtifactsListParams" }
+      }
+    },
+    "ArtifactsListOkResponseFrame": {
+      "type": "object",
+      "additionalProperties": false,
+      "required": ["type", "id", "ok", "payload"],
+      "properties": {
+        "type": { "const": "res" },
+        "id": { "type": "string" },
+        "ok": { "const": true },
+        "payload": { "$ref": "#/$defs/ArtifactsListResult" },
+        "error": { "type": "null" }
+      }
+    },
+    "ArtifactsListErrorResponseFrame": {
+      "type": "object",
+      "additionalProperties": false,
+      "required": ["type", "id", "ok", "error"],
+      "properties": {
+        "type": { "const": "res" },
+        "id": { "type": "string" },
+        "ok": { "const": false },
+        "payload": { "type": "null" },
+        "error": { "$ref": "#/$defs/RpcError" }
+      }
+    },
+    "ArtifactsListResponseFrame": {
+      "oneOf": [
+        { "$ref": "#/$defs/ArtifactsListOkResponseFrame" },
+        { "$ref": "#/$defs/ArtifactsListErrorResponseFrame" }
+      ]
+    }
+  }
+}
